+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_moving_to_variable/visitor.dart
@@ -61,6 +61,7 @@ class _BlockVisitor extends RecursiveAstVisitor<void> {
   @override
   void visitMethodInvocation(MethodInvocation node) {
     if (node.parent is CascadeExpression ||
+        node.parent is VariableDeclaration ||
         (node.staticType?.isVoid ?? false)) {
       return;
     }