+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_moving_to_variable/examples/assignment_example.dart
@@ -1,0 +1,8 @@
+void main() {
+  final var1 = someFunction();
+  final var2 = someFunction();
+}
+
+String someFunction() {
+  return 'qwe';
+}